app/requirements.txt
psycopg2-binary
Flask-SQLAlchemy
flask-cors
pytest
python-dotenv
flask-restx
gunicorn
requests
inquirer
